requirements.txt
aniso8601==7.0.0
asgiref==3.2.5
autopep8==1.5.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-filter==2.2.0
django-graphiql==0.4.4
django-graphql-jwt==0.3.1
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.1
idna==2.9
mysqlclient==1.4.5
Pillow==7.0.0
promise==2.3
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
